PR Review: remove testing base class
