+++ corehq/apps/sso/tests/test_context_helpers.py
@@ -1,13 +1,13 @@
 import datetime
 from unittest.mock import patch
 
-from django.test import SimpleTestCase
+from django.test import SimpleTestCase, TestCase
 
 from corehq.apps.sso.utils.context_helpers import (
     render_multiple_to_strings,
     get_idp_cert_expiration_email_context,
 )
-from corehq.apps.sso.tests.base_test import BaseIdPTest
+from corehq.apps.sso.tests import generator
 
 
 class TestSimpleIdPContextHelpers(SimpleTestCase):
@@ -22,10 +22,28 @@ class TestSimpleIdPContextHelpers(SimpleTestCase):
             mock_render.reset_mock()
 
 
-class TestIdPContextHelpers(BaseIdPTest):
+class TestIdPContextHelpers(TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.account = generator.get_billing_account_for_idp()
+
+    def setUp(self):
+        super().setUp()
+        self.idp = generator.create_idp('vaultwax', self.account)
 
     def test_get_idp_cert_expiration_email_context(self):
         self.idp.date_idp_cert_expiration = datetime.datetime.utcnow()
         self.idp.save()
         self.assertSetEqual(set(get_idp_cert_expiration_email_context(self.idp)),
                             {"subject", "from", "to", "bcc", "html", "plaintext"})
+
+    def tearDown(self):
+        self.idp.delete()
+        super().tearDown()
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.account.delete()
+        super().tearDownClass()
